+++ ZeroLibraries/Common/String/StringBuilder.hpp
@@ -42,7 +42,7 @@ public:
   void Append(char character);
   void AppendFormat(cstr format, ...);
   void Append(Rune rune);
-  void Append(cstr begin, uint sizeInBytes);
+  void Append(cstr begin, size_t sizeInBytes);
   char& operator[](size_t index);
 
   void Repeat(size_t count, StringParam str);
@@ -102,7 +102,7 @@ inline StringBuilder& operator<<(StringBuilder& builder, const type& value)
 {
   const uint bufferSize = 128;
   char buffer[bufferSize];
-  uint size = ToBuffer(buffer, bufferSize, value);
+  size_t size = ToBuffer(buffer, bufferSize, value);
   builder.Append(String(buffer, buffer+size));
   return builder;
 }
@@ -114,10 +114,10 @@ inline String GetBinaryString(const T& value)
   byte* valueCursor = (byte*)&value;
 
   // Read every byte (from left to right)
-  for (uint i = 0; i < sizeof(value); ++i)
+  for (size_t i = 0; i < sizeof(value); ++i)
   {
     // Read every bit in the byte (from left to right)
-    for (uint j = 0; j < 8; ++j)
+    for (size_t j = 0; j < 8; ++j)
       result += *valueCursor & LBIT(j) ? '1' : '0';
 
     // Next byte
